refactor: reorganize components and clean up unused code
- Move CreateTermModal from Routes to components directory for better organization
- Remove unused App.css file and clean up MainHeader navigation links
- Fix code formatting and remove redundant imports in various files
