+++ client/src/Routes/AdminHomePage.jsx
@@ -3,6 +3,7 @@ import axios from "../api/axios";
 import ReportCard from "../components/ReportCard";
 import { toast } from "react-toastify";
 import Screen from "../components/Screen";
+import CreateTermModal from "../components/CreateTermModal";
 
 export default function AdminHomePage() {
   const [reports, setReports] = useState([]);
@@ -44,32 +45,33 @@ export default function AdminHomePage() {
       toast.error("حدث خطأ أثناء رفض البلاغ");
     }
   }
-  if (reports.length === 0) {
-    return (
-      <Screen
-        title="Banned Accounts"
-        className="p-2 sm:p-4 md:p-6 flex items-center justify-center"
-      >
-        <div className="text-red-400 text-2xl">لا يوجد بلاغات</div>
-      </Screen>
-    );
-  }
+
   return (
-    <div className="min-h-screen max-h-max w-full bg-gradient-to-b from-TAF-200 via-white to-TAF-200">
-      <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-2 mx-4 p-4">
-        {reports.map((report) => (
-          <ReportCard
-            key={report.reportId}
-            reportId={report.reportId}
-            reason={report.content}
-            comment={report.comment}
-            handleReject={handleReject}
-            commentWriter={report.comment.authorName}
-            onReject={() => onReject(report.reportId)}
-            onDeleteComment={() => handleDeleteReport(report.reportId)}
-          />
-        ))}
+    <Screen>
+      <div
+        className={`${
+          reports.length > 0
+            ? "grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-2 mx-4 p-4"
+            : "flex items-center justify-center h-screen"
+        }`}
+      >
+        {reports.length > 0 ? (
+          reports.map((report) => (
+            <ReportCard
+              key={report.reportId}
+              reportId={report.reportId}
+              reason={report.content}
+              comment={report.comment}
+              handleReject={handleReject}
+              commentWriter={report.comment.authorName}
+              onReject={() => onReject(report.reportId)}
+              onDeleteComment={() => handleDeleteReport(report.reportId)}
+            />
+          ))
+        ) : (
+          <div className="text-red-400 text-2xl">لا يوجد بلاغات</div>
+        )}
       </div>
-    </div>
+    </Screen>
   );
 }
