+++ client/src/components/RootLayoutComponents/MainHeader.jsx
@@ -100,10 +100,6 @@ export default function MainHeader() {
                   linkTo={"التعليقات المخفية"}
                   route={"/admin/hiddencomments"}
                 />
-                <NavigationLink
-                  linkTo={"إنشاء ترم جديد"}
-                  route={"/admin/createTerm"}
-                />
               </>
             )}
           </div>
